+++ spec/display_manager_spec.rb
@@ -8,8 +8,10 @@ RSpec.describe ArduinoCI::DisplayManager do
   end
 
   context "with_display" do
-    it "Properly enables and disables" do
-      manager = ArduinoCI::DisplayManager::instance
+    manager = ArduinoCI::DisplayManager::instance
+    manager.disable
+
+    it "Properly enables and disables when not previously enabled" do
       expect(manager.enabled).to be false
       manager.with_display do |environment|
         expect(manager.enabled).to be true
@@ -18,6 +20,18 @@ RSpec.describe ArduinoCI::DisplayManager do
         expect(also_manager.enabled).to be true
       end
       expect(manager.enabled).to be false
+    end
+
+    it "Properly enables and disables when previously enabled" do
+      manager.enable
+      expect(manager.enabled).to be true
+      manager.with_display do |environment|
+        expect(manager.enabled).to be true
+        expect(environment.class).to eq(Hash)
+        also_manager = ArduinoCI::DisplayManager::instance
+        expect(also_manager.enabled).to be true
+      end
+      expect(manager.enabled).to be true
     end
   end
 end