+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `DisplayManager` class for managing Xvfb instance if needed
 
 ### Changed
+- `DisplayManger.with_display` doesn't `disable` if the display was enabled prior to starting the block
 
 ### Deprecated